requirements.txt
pymongo
python-dotenv
dnspython
fastapi
uvicorn[standard]